update file writer namepsace
