+++ src/Core/File/Writer/FileWriterData.php
@@ -24,9 +24,46 @@
  * International Registered Trademark & Property of PrestaShop SA
  */
 
-namespace PrestaShop\PrestaShop\Core\Export\Exception;
+namespace PrestaShop\PrestaShop\Core\File\Writer;
 
-class FileWritingException extends ExportException
+/**
+ * Class FileWriterData stores data that should be written to export file
+ */
+final class FileWriterData implements FileWriterDataInterface
 {
-    const CANNOT_OPEN_FILE_FOR_WRITING = 10;
+    /**
+     * @var array
+     */
+    private $headers;
+
+    /**
+     * @var array
+     */
+    private $rows;
+
+    /**
+     * @param array $headers
+     * @param array $rows
+     */
+    public function __construct(array $headers, array $rows)
+    {
+        $this->headers = $headers;
+        $this->rows = $rows;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getHeaders()
+    {
+        return $this->headers;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getRows()
+    {
+        return $this->rows;
+    }
 }
