+++ src/PrestaShopBundle/Controller/Admin/Configure/AdvancedParameters/SqlRequestController.php
@@ -41,8 +41,7 @@ use PrestaShop\PrestaShop\Core\Domain\SqlManagement\Query\GetSqlRequestSettings;
 use PrestaShop\PrestaShop\Core\Domain\SqlManagement\SqlRequestExecutionResult;
 use PrestaShop\PrestaShop\Core\Domain\SqlManagement\SqlRequestSettings;
 use PrestaShop\PrestaShop\Core\Domain\SqlManagement\ValueObject\SqlRequestId;
-use PrestaShop\PrestaShop\Core\Export\Exception\ExportException;
-use PrestaShop\PrestaShop\Core\Export\Exception\FileWritingException;
+use PrestaShop\PrestaShop\Core\File\Writer\Exception\FileWritingException;
 use PrestaShop\PrestaShop\Core\Form\FormHandlerInterface;
 use PrestaShop\PrestaShop\Core\Search\Filters\RequestSqlFilters;
 use PrestaShopBundle\Controller\Admin\FrameworkBundleAdminController;
