+++ src/PrestaShopBundle/Resources/config/services/core/file.yml
@@ -14,3 +14,8 @@ services:
         class: 'PrestaShop\PrestaShop\Core\Proxy\CachedFileFinderProxy'
         arguments:
             - '@prestashop.core.file.finder.invoice_model'
+
+    prestashop.core.file.writer.csv_file_writer:
+        class: 'PrestaShop\PrestaShop\Core\File\Writer\CsvFileWriter'
+        arguments:
+            - '@prestashop.core.export.dir'
